requirements.txt
numpy~=2.2.1
matplotlib~=3.10.0
scipy~=1.14.1
pygame~=2.6.1
pyyaml~=6.0.2